+++ test/measure_test.dart
@@ -6,7 +6,7 @@ void main() {
     final measure = Measure.fromNoteList([
       Note(1, 2),
     ]);
-    expect(measure.notes[0].timing, equals(const Timing(NoteType.whole, 1)));
+    expect(measure.notes[0].timing, equals(Timing.ofWholeNote(1)));
   });
 
   test('Measure.fromNoteList calculates half note timing', () {
@@ -14,8 +14,8 @@ void main() {
       Note(1, 2),
       Note(1, 2),
     ]);
-    expect(measure.notes[0].timing, equals(const Timing(NoteType.half, 1)));
-    expect(measure.notes[1].timing, equals(const Timing(NoteType.half, 2)));
+    expect(measure.notes[0].timing, equals(Timing.ofHalfNote(1)));
+    expect(measure.notes[1].timing, equals(Timing.ofHalfNote(2)));
   });
 
   test('Measure.fromNoteList calculates quarter note timing', () {
@@ -25,10 +25,10 @@ void main() {
       Note(3, 4),
       Note(5, 6),
     ]);
-    expect(measure.notes[0].timing, equals(const Timing(NoteType.quarter, 1)));
-    expect(measure.notes[1].timing, equals(const Timing(NoteType.quarter, 2)));
-    expect(measure.notes[2].timing, equals(const Timing(NoteType.quarter, 3)));
-    expect(measure.notes[3].timing, equals(const Timing(NoteType.quarter, 4)));
+    expect(measure.notes[0].timing, equals(Timing.ofQuarterNote(1)));
+    expect(measure.notes[1].timing, equals(Timing.ofQuarterNote(2)));
+    expect(measure.notes[2].timing, equals(Timing.ofQuarterNote(3)));
+    expect(measure.notes[3].timing, equals(Timing.ofQuarterNote(4)));
   });
 
   test('Measure.fromNoteList calculates eighth note timing', () {
@@ -42,14 +42,14 @@ void main() {
       Note(3, 4),
       Note(5, 6),
     ]);
-    expect(measure.notes[0].timing, equals(const Timing(NoteType.eighth, 1)));
-    expect(measure.notes[1].timing, equals(const Timing(NoteType.eighth, 2)));
-    expect(measure.notes[2].timing, equals(const Timing(NoteType.eighth, 3)));
-    expect(measure.notes[3].timing, equals(const Timing(NoteType.eighth, 4)));
-    expect(measure.notes[4].timing, equals(const Timing(NoteType.eighth, 5)));
-    expect(measure.notes[5].timing, equals(const Timing(NoteType.eighth, 6)));
-    expect(measure.notes[6].timing, equals(const Timing(NoteType.eighth, 7)));
-    expect(measure.notes[7].timing, equals(const Timing(NoteType.eighth, 8)));
+    expect(measure.notes[0].timing, equals(Timing.ofEighthNote(1)));
+    expect(measure.notes[1].timing, equals(Timing.ofEighthNote(2)));
+    expect(measure.notes[2].timing, equals(Timing.ofEighthNote(3)));
+    expect(measure.notes[3].timing, equals(Timing.ofEighthNote(4)));
+    expect(measure.notes[4].timing, equals(Timing.ofEighthNote(5)));
+    expect(measure.notes[5].timing, equals(Timing.ofEighthNote(6)));
+    expect(measure.notes[6].timing, equals(Timing.ofEighthNote(7)));
+    expect(measure.notes[7].timing, equals(Timing.ofEighthNote(8)));
   });
 
   test('Measure.fromNoteList calculates sixteenth note timing', () {
@@ -71,38 +71,22 @@ void main() {
       Note(3, 4),
       Note(5, 6),
     ]);
-    expect(
-        measure.notes[0].timing, equals(const Timing(NoteType.sixteenth, 1)));
-    expect(
-        measure.notes[1].timing, equals(const Timing(NoteType.sixteenth, 2)));
-    expect(
-        measure.notes[2].timing, equals(const Timing(NoteType.sixteenth, 3)));
-    expect(
-        measure.notes[3].timing, equals(const Timing(NoteType.sixteenth, 4)));
-    expect(
-        measure.notes[4].timing, equals(const Timing(NoteType.sixteenth, 5)));
-    expect(
-        measure.notes[5].timing, equals(const Timing(NoteType.sixteenth, 6)));
-    expect(
-        measure.notes[6].timing, equals(const Timing(NoteType.sixteenth, 7)));
-    expect(
-        measure.notes[7].timing, equals(const Timing(NoteType.sixteenth, 8)));
-    expect(
-        measure.notes[8].timing, equals(const Timing(NoteType.sixteenth, 9)));
-    expect(
-        measure.notes[9].timing, equals(const Timing(NoteType.sixteenth, 10)));
-    expect(
-        measure.notes[10].timing, equals(const Timing(NoteType.sixteenth, 11)));
-    expect(
-        measure.notes[11].timing, equals(const Timing(NoteType.sixteenth, 12)));
-    expect(
-        measure.notes[12].timing, equals(const Timing(NoteType.sixteenth, 13)));
-    expect(
-        measure.notes[13].timing, equals(const Timing(NoteType.sixteenth, 14)));
-    expect(
-        measure.notes[14].timing, equals(const Timing(NoteType.sixteenth, 15)));
-    expect(
-        measure.notes[15].timing, equals(const Timing(NoteType.sixteenth, 16)));
+    expect(measure.notes[0].timing, equals(Timing.ofSixteenthNote(1)));
+    expect(measure.notes[1].timing, equals(Timing.ofSixteenthNote(2)));
+    expect(measure.notes[2].timing, equals(Timing.ofSixteenthNote(3)));
+    expect(measure.notes[3].timing, equals(Timing.ofSixteenthNote(4)));
+    expect(measure.notes[4].timing, equals(Timing.ofSixteenthNote(5)));
+    expect(measure.notes[5].timing, equals(Timing.ofSixteenthNote(6)));
+    expect(measure.notes[6].timing, equals(Timing.ofSixteenthNote(7)));
+    expect(measure.notes[7].timing, equals(Timing.ofSixteenthNote(8)));
+    expect(measure.notes[8].timing, equals(Timing.ofSixteenthNote(9)));
+    expect(measure.notes[9].timing, equals(Timing.ofSixteenthNote(10)));
+    expect(measure.notes[10].timing, equals(Timing.ofSixteenthNote(11)));
+    expect(measure.notes[11].timing, equals(Timing.ofSixteenthNote(12)));
+    expect(measure.notes[12].timing, equals(Timing.ofSixteenthNote(13)));
+    expect(measure.notes[13].timing, equals(Timing.ofSixteenthNote(14)));
+    expect(measure.notes[14].timing, equals(Timing.ofSixteenthNote(15)));
+    expect(measure.notes[15].timing, equals(Timing.ofSixteenthNote(16)));
   });
 
   test('Measure.fromNoteList syncs calculated timing with concurrent notes',
@@ -113,11 +97,11 @@ void main() {
       Note(3, 4),
       Note(5, 6),
     ]);
-    expect(measure.notes[0].timing, equals(const Timing(NoteType.quarter, 1)));
+    expect(measure.notes[0].timing, equals(Timing.ofQuarterNote(1)));
     expect(measure.notes[0].timing, equals(measure.notes[0].and!.timing));
     expect(measure.notes[0].timing, equals(measure.notes[0].and!.and!.timing));
-    expect(measure.notes[1].timing, equals(const Timing(NoteType.quarter, 2)));
-    expect(measure.notes[2].timing, equals(const Timing(NoteType.quarter, 3)));
-    expect(measure.notes[3].timing, equals(const Timing(NoteType.quarter, 4)));
+    expect(measure.notes[1].timing, equals(Timing.ofQuarterNote(2)));
+    expect(measure.notes[2].timing, equals(Timing.ofQuarterNote(3)));
+    expect(measure.notes[3].timing, equals(Timing.ofQuarterNote(4)));
   });
 }